ml_engine/requirements.txt
#annotated-doc==0.0.3
#annotated-types==0.7.0
#anyio==4.11.0
fastapi==0.120.4
#idna==3.11
#pydantic==2.12.3
#pydantic_core==2.41.4
#sniffio==1.3.1
#starlette==0.49.1
#typing-inspection==0.4.2
#typing_extensions==4.15.0
uvicorn==0.38.0
requests==2.32.5
datetime==5.5
#typing==3.7.4.3
python-dateutil==2.9.0
pandas==2.3.3
numpy==2.3.4
lightgbm==4.6.0
asyncio==4.0.0
httpx==0.28.1